# lib/api-zod/src/generated/types/globalLeaderboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Fire Tournament Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { GlobalPlayerRank } from "./globalPlayerRank";

export interface GlobalLeaderboardResponse {
  players: GlobalPlayerRank[];
  total: number;
  page: number;
  limit: number;
}