# lib/api-zod/src/generated/types/globalPlayerRank.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Fire Tournament Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface GlobalPlayerRank {
  rank: number;
  user: User;
  totalPoints: number;
  totalKills: number;
  matchesPlayed: number;
  totalEarnings: number;
}