# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Fire Tournament Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  freeFireUid: string;
  inGameName: string;
  whatsappNumber: string;
  isAdmin: boolean;
  totalEarnings: number;
  globalRank: number;
  matchesPlayed: number;
  createdAt: Date;
}